Includes Error for gap width and measures the profile of samples
New file is to use the ECF function on a pixel by pixel level to extract the changes in width vertically within a sample. Changes to OutputFigure are to include this option.
FittedEdges has been updated to give a gap measurement that has an associated standard error and to now include a measure of any gap between the sample and the platen that illustrates if the sample is in contact or not.