Change to be explicit about frame distribution in time
video duration information is automatically calculated from the number of recorded frames and the frame rate that is set by the user starting the Thor cam recording. This does not take into account dropped frames. In the unfortunate case where the computer is unable to record every frame collected by the camera the number of frames will be less than expected and the duration of the video will appear shorter than the real time it took to record.
This change will allow for manual correction later by only changing the framerate to match the recorded framerate instead of the user requested framerate